requirements.txt
flask
flask_cors
numpy
opencv-python-headless
pillow
face_recognition
dlib
requests